Squashed commit of the following:
commit 6c2474b42b41b6c630e0a12490610552cab7a0b0
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 22:09:12 2026 -0600

    fixes

commit d45bd745008ad30f1f93d4e78f93bd123b904d41
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 21:22:04 2026 -0600

    tests+shims

commit 7689f3c35e37809897a667929767c53224563ef3
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 21:05:04 2026 -0600

    Update README.md

commit 8b98b4acb07171f7f259b0db91aea6961d5b0a8f
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 20:42:28 2026 -0600

    compile fixes

commit 42c4354015458276bdc0202ba8e3413d03709b2a
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 19:04:03 2026 -0600

    tests

commit fb6acb901523d0fd1e7cfa0d6c3015dcb7f41b8b
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 18:33:44 2026 -0600

    tests

commit 2ab33a35d16ef00fa9158d1157e57d9479606b1e
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 18:02:30 2026 -0600

    fixes

commit 6d3ae6363bfc9b9c58574da3fe682c6143fb1f31
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 17:10:26 2026 -0600

    fixes

commit 2ee7cec57e1b84b286b984a246d4d5f77459199d
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 16:37:30 2026 -0600

    fixes

commit 8f5c022c4571d8a4640964cb607111304de9a6aa
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 13:19:08 2026 -0600

    fixes

commit d2744ed8672499bd2b1b4eeb76c7469003086ebf
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 13:01:31 2026 -0600

    Update Helioduino.cpp

commit 21bf82875943f18c53fd7627c46f30f387c55737
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 11:55:35 2026 -0600

    fractional timezone

commit cfd8915895b9263ac6f8996ba33a40c919c52cd4
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Wed Sep 2 10:03:55 2026 -0600

    Update HelioInputDrivers.h

commit 8b98513ae77335d268843df0fef49512d5fa3c8e
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Tue Sep 1 20:42:45 2026 -0600

    fixes

commit dc20324a44a1b310b86c0659f643c8672274e9e7
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Tue Sep 1 18:19:01 2026 -0600

    Update Helioduino.cpp

commit 4139fc9e63efc6bfba385f42efbf411649bce9bf
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Tue Sep 1 17:28:18 2026 -0600

    Update README.md

commit 56ab6b4a33a7f1cca705f5b9141785054b5a0c60
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Tue Sep 1 15:39:38 2026 -0600

    Update Helioduino.cpp

commit 8ff21599c4a2af447c55d5e11551eeb4d01a0666
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Tue Sep 1 15:27:55 2026 -0600

    Update README.md

commit 41ec11eea8b9f5017bb9e341f3c5c672c8a9e64f
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Tue Sep 1 14:59:33 2026 -0600

    updates

commit 00b3dc79acc5df2496a57906deead5c020bfabcc
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Mon Aug 31 23:18:34 2026 -0600

    Update HelioStrings.h

commit ce0385833b06d3b51b7203352e6bb6b7b53526bf
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Mon Aug 31 21:44:00 2026 -0600

    Update HelioScheduler.cpp

commit b39441b791bd7b01a15bd607cf6d8d910c16ecf3
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Mon Aug 31 13:53:41 2026 -0600

    fixes

commit ff0fa9921d892773ece518be5ade96c188e0b47f
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Fri Aug 28 22:00:49 2026 -0600

    Update HelioCoreLogic.h

commit 6aff2fb0ed7722382af665bc9bc2b8fa823d2f8e
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Fri Aug 28 15:35:40 2026 -0600

    removing validate crap

commit 09bc6dd6efced1027ef12fd552c03c3702ff1522
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Tue Aug 25 13:32:14 2026 -0600

    Update HelioAttachments.hpp

commit 7dd07d565db389ef44815bf503ceae8f8b5d3973
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Sat Aug 22 16:22:25 2026 -0600

    Update HelioInterfaces.hpp

commit fd04b7c1656ba953168b68e531620014f6ba415f
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Sat Aug 22 15:28:34 2026 -0600

    Update HelioCallback.hh

commit f8fc9e32581fb9b7f5b21601169622b0300603b1
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Sat Aug 22 13:50:45 2026 -0600

    fixing main hpp

commit 56a2240651599c967ed4e804240008e091fe80ed
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Fri Aug 21 00:29:00 2026 -0600

    Update HelioModules.cpp

commit 69d7b06281d1ba643cb4e0fcfe89741bd2bb226b
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Thu Aug 20 21:37:47 2026 -0600

    upping version

commit 689b3dd6d70a7ffe0a9c0119a02c3ef18d68a3aa
Author: NachtRaveVL <NachtRaveVL@users.noreply.github.com>
Date:   Thu Aug 20 19:34:15 2026 -0600

    Removing old stuff & consistency

+++ src/HelioStreams.h
@@ -36,7 +36,7 @@ public:
 
 protected:
     I2C_eeprom *_eeprom;
-    uint16_t _readAddress, _writeAddress, _endAddress;
+    uint32_t _readAddress, _writeAddress, _endAddress;
 };
 
 
@@ -63,7 +63,7 @@ protected:
 
 class HelioWiFiStorageFileStream : public Stream {
 public:
-    HelioWiFiStorageFileStream(WiFiStorageFile file, uintptr_t seekPos = 0);
+    HelioWiFiStorageFileStream(WiFiStorageFile file, uint32_t seekPos = 0);
     virtual ~HelioWiFiStorageFileStream();
 
     virtual int available() override;
@@ -81,9 +81,9 @@ protected:
     WiFiStorageFile _file;
     uint8_t _buffer[HELIO_WIFISTREAM_BUFFER_SIZE];
     size_t _bufferOffset;
-    uintptr_t _bufferFileOffset;
+    uint32_t _bufferFileOffset;
     WiFiStorageFileDirection _bufferDirection;
-    uintptr_t _readOffset, _writeOffset, _endOffset;
+    uint32_t _readOffset, _writeOffset, _endOffset;
 
     void prepareReadBuffer();
     void prepareWriteBuffer();
